Refactor ElasticsearchPresenter
